ensure Remove option lists only installed packages
NOTE: I'm comparing with the original code here and... I'm very confused what the packagelist file is for. It looks like it's supposed to hold a cache of *remote* package names, but then it's used for querying local packages? Why would you cache something local and not remote? And it doesn't seem to work properly for local packages anyway, so... I'm just gonna replace that bit with something I understand lol